+++ src/js/bootstrapValidator.js
@@ -229,7 +229,11 @@
                 total     = fields.length,
                 type      = fields.attr('type'),
                 updateAll = (total === 1) || ('radio' === type) || ('checkbox' === type),
-                event     = ('radio' === type || 'checkbox' === type || 'file' === type || 'SELECT' === fields.eq(0).get(0).tagName) ? 'change' : this._changeEvent;
+                event     = ('radio' === type || 'checkbox' === type || 'file' === type || 'SELECT' === fields.eq(0).get(0).tagName) ? 'change' : this._changeEvent,
+                trigger   = (this.options.fields[field].trigger || this.options.trigger || event).split(' '),
+                events    = $.map(trigger, function(item) {
+                    return item + '.update.bv';
+                }).join(' ');
 
             for (var i = 0; i < total; i++) {
                 var $field    = fields.eq(i),
@@ -248,7 +252,7 @@
                 $parent.find('i[data-bv-icon-for="' + field + '"]').remove();
 
                 // Whenever the user change the field value, mark it as not validated yet
-                $field.off(event + '.update.bv').on(event + '.update.bv', function() {
+                $field.off(events).on(events, function() {
                     // Reset the flag
                     that._submitIfValid = false;
                     that.updateStatus($(this), that.STATUS_NOT_VALIDATED);
@@ -315,10 +319,9 @@
             }
 
             // Set live mode
-            var trigger = this.options.fields[field].trigger || this.options.trigger || event,
-                events  = $.map(trigger.split(' '), function(item) {
-                    return item + '.live.bv';
-                }).join(' ');
+            events = $.map(trigger, function(item) {
+                return item + '.live.bv';
+            }).join(' ');
             switch (this.options.live) {
                 case 'submitted':
                     break;